NSLog format update
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/branches/gnustep_testplant_branch@40244 72102866-910b-0410-8b05-ffd578937521

+++ Source/GSHorizontalTypesetter.m
@@ -1234,7 +1234,7 @@ restart: ;
       static NSUInteger RecursionCount = 0;
       RecursionCount++;
       if (RecursionCount > 1)
-        NSLog(@"%s:entering %d level recursion", __PRETTY_FUNCTION__, RecursionCount);
+        NSLog(@"%s:entering %ld level recursion", __PRETTY_FUNCTION__, (long)RecursionCount);
 
       GSHorizontalTypesetter *temp;
